docs: header recipes
